+++ inference/run_predict_full_rnd_gnns.sh
@@ -11,7 +11,7 @@ for i in {01..05}; do
     OUTPUT_SUBDIR="${OUTPUT_DIR}/test_gnn_graph-hnet-pseudo_baseline_${i}"
 
     # 运行 Python 脚本
-    python inference/predict_full_patched_gnn.py \
+    python inference/predict_full_patched_gnns.py \
         --img_path "$INPUT_PATH" \
         --patch_encoder_path "$PATCH_ENCODER_PATH" \
         --gnn_path "$GNN_PATH" \
